requirements.txt
discord.py>=2.4.0
wavelink>=3.3.0
aiohttp>=3.9.0